# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FonAd / DAADD API
  version: 1.0.0
  description: |
    FonAd (DAADD) is a full-featured digital advertising and analytics platform.
    It provides campaign management, real-time event tracking, AI-driven optimization,
    anomaly detection, cross-device attribution, benchmarking, storytelling reports,
    viewer reward programs, team collaboration, and notification services.
  contact:
    name: FonAd Engineering
    email: engineering@fonad.io
servers:
  - url: http://localhost:4000/api/v1
    description: Local development server
security:
  - BearerAuth: []
...
paths:
...
  /rewards:
    get:
      tags: [Rewards]
      summary: Get user rewards
      description: Returns a paginated list of rewards for the authenticated user.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Order'
      responses:
        '200':
          description: Paginated list of rewards
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Reward'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /rewards/claim/{eventId}:
    post:
      tags: [Rewards]
      summary: Claim a reward
      description: Claims a reward for a specific ad event. The reward value is determined by the campaign's reward_value setting.
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ad event UUID to claim a reward for
      responses:
        '201':
          description: Reward claimed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Reward'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /rewards/balance:
    get:
      tags: [Rewards]
      summary: Get reward balance
      description: Returns the total reward balance for the authenticated user.
      responses:
        '200':
          description: Reward balance
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          balance:
                            type: number
                            format: float
                            description: Total available reward balance
                          total_earned:
                            type: number
                            format: float
                            description: Lifetime total rewards earned
                          total_redeemed:
                            type: number
                            format: float
                            description: Lifetime total rewards redeemed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  # ─────────────────────────── Teams ───────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/register
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
    Sort:
      name: sort
      in: query
      schema:
        type: string
        default: created_at
      description: Field name to sort by
    Order:
      name: order
      in: query
      schema:
        type: string
        enum: [ASC, DESC]
        default: DESC
      description: Sort order
...
  responses:
    BadRequest:
      description: Bad request -- missing or invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized -- missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    RewardStatus:
      type: string
      enum: [pending, credited, redeemed, expired]
...
    Reward:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        campaign_id:
          type: string
          format: uuid
        event_id:
          type: string
          format: uuid
          nullable: true
        amount:
          type: number
          format: float
          description: Reward amount in USD
        status:
          $ref: '#/components/schemas/RewardStatus'
        earned_at:
          type: string
          format: date-time
        redeemed_at:
          type: string
          format: date-time
          nullable: true
      required: [id, user_id, campaign_id, amount, status, earned_at]
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          description: Response payload (type varies by endpoint)
      required: [success]
    PaginatedResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items: {}
        pagination:
          type: object
          properties:
            total:
              type: integer
              description: Total number of records
            page:
              type: integer
              description: Current page number
            limit:
              type: integer
              description: Items per page
            totalPages:
              type: integer
              description: Total number of pages
            hasNext:
              type: boolean
            hasPrev:
              type: boolean
          required: [total, page, limit, totalPages, hasNext, hasPrev]
      required: [success, data, pagination]
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
              description: Human-readable error message
            code:
              type: string
              description: Machine-readable error code
              enum:
                - BAD_REQUEST
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - INTERNAL_ERROR
                - INVALID_CREDENTIALS
                - EMAIL_EXISTS
                - ACCOUNT_INACTIVE
                - INVALID_TRANSITION
                - INVALID_BUDGET
                - AI_NOT_READY
            details:
              description: Additional error context
          required: [message, code]
      required: [success, error]